AgainB

+++ 2019.2.25/AgainB.c
@@ -9,7 +9,7 @@ using namespace std;
 const int N = 1e5 + 5;
 int n;
 int mark[N];
-vector <int> aux[N];
+vector <ll> aux[N];
 int main()
 {
 	mmset(mark,0);
@@ -21,13 +21,13 @@ int main()
 		scanf("%d",&num);
 		aux[num].push_back(i);
 	}
-	int res = 0;
+	ll res = 0;
 	for(int i = 1; i < n / 2; i++)
 	{
 		res += min(fabs(aux[i][0] - aux[i+1][0]) + fabs(aux[i][1] - aux[i+1][1]),fabs(aux[i][1] - aux[i+1][0]) + fabs(aux[i][0] - aux[i+1][1]));
 	}
 	res += (aux[1][0] + aux[1][1] - 2);
-	printf("%d\n",res);
+	printf("%lld\n",res);
 
 	return 0;
 }
